Refactored Backend
Made the backend more modular and efficient
