+++ backend/app/templates/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>AllDone App</title>
+</head>
+
+<body>
+    <h1>AllDone App</h1>
+</body>
+
+</html>
