--- backend/templates/index.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>AllDone App</title>
-    <link rel="stylesheet" href="{{url_for('static', filename='css/style.css')}}">
-</head>
-
-<body>
-    <h1>AllDone App</h1>
-    <form action="{{url_for('home')}}" class="input-container" method="POST">
-        <input type="text" placeholder="Enter Task" id="input-task" name="task_name">
-        <button id="add-task">+</button>
-    </form>
-    <h2>Tasks</h2>
-    <div class="task-container">
-        {%for item in items %}
-        <div class="task">
-            <p>{{ item.name}}</p>
-            <div class="actions">
-                <form action="{{url_for('checked_task', task_id=item.id)}}" method="POST">
-                    <input type="checkbox" {% if item.checked %}checked{% endif %} onchange="this.form.submit()">
-                </form>
-                <form action="{{url_for('delete_task', task_id=item.id)}}" method="POST">
-                    <button>Delete</button>
-                </form>
-            </div>
-        </div>
-        {% endfor %}
-    </div>
-</body>
-
-</html>
