+++ backend/tests/conftest.py
@@ -1,4 +1,9 @@
-import sys
-import os
+import pytest
+from app import create_app
 
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../")))
+@pytest.fixture
+def client():
+    flask_app = create_app()
+    flask_app.config['TESTING'] = True
+    with flask_app.test_client() as client:
+        yield client
